Revert back to old js files
